コントローラーをまとめ、コードを修正しました。
