--- src/Form/Article1Type.php
@@ -1,27 +1,0 @@
-<?php
-
-namespace App\Form;
-
-use App\Entity\Article;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class Article1Type extends AbstractType
-{
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add('title')
-            ->add('content')
-            ->add('created_at')
-        ;
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults([
-            'data_class' => Article::class,
-        ]);
-    }
-}
