+++ templates/article/_delete_form.html.twig
@@ -1,5 +1,5 @@
-<form method="post" action="{{ path('article_delete', {'id': article.id}) }}" onsubmit="return confirm('Are you sure you want to delete this item?');">
+<form method="post" action="{{ path('article_delete', {'id': article.id}) }}" onsubmit="return confirm('削除しますか？');">
     <input type="hidden" name="_method" value="DELETE">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ article.id) }}">
-    <button class="btn">Delete</button>
+    <input type="hidden" name="_delete_token" value="{{ csrf_token('article_delete' ~ article.id) }}">
+    <button class="btn">削除する</button>
 </form>
