+++ templates/article/_form.html.twig
@@ -1,4 +1,3 @@
 {{ form_start(form) }}
-    {{ form_widget(form) }}
-    <button class="btn">{{ button_label|default('Save') }}</button>
+{{ form_widget(form) }}
 {{ form_end(form) }}
