+++ templates/article/edit.html.twig
@@ -1,11 +1,11 @@
 {% extends 'base.html.twig' %}
 
-{% block title %}Edit Article{% endblock %}
+{% block title %}記事編集{% endblock %}
 
 {% block body %}
-    <h1>Edit Article</h1>
+    <h1>記事編集</h1>
 
-    {{ include('article/_form.html.twig', {'button_label': 'Update'}) }}
+    {{ include('article/_form.html.twig') }}
 
     <a href="{{ path('article_index') }}">back to list</a>
