+++ templates/base.html.twig
@@ -20,17 +20,12 @@
     <body>
         <div class="header">
             <nav class="navbar navbar-dark bg-dark mt-3 mb-3 justify-content-center">
-                <a href="/home" class="navbar-brand">
+                <a href="{{ path('article_index') }}" class="navbar-brand">
                     ブログ
                 </a>
                 <ul class="navbar-nav mr-2">
                     <li class="nav-item">
-                        <a class="nav-link" href="/post">記事投稿</a>
-                    </li>
-                </ul>
-                <ul class="navbar-nav mr-2">
-                    <li class="nav-item">
-                        <a class="nav-link" href="/update">記事編集</a>
+                        <a class="nav-link" href="{{ path('article_new') }}">記事投稿</a>
                     </li>
                 </ul>
             </nav>
