--- templates/delete/index.html.twig
@@ -1,10 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}更新{% endblock %}
-
-{% block body %}
-    <h2>記事削除</h2>
-    {{ form_start(form) }}
-    {{ form_widget(form) }}
-    {{ form_end(form) }}
-{% endblock %}
