--- templates/home/find.html.twig
@@ -1,8 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}{{ data.title }}{% endblock %}
-
-{% block body %}
-    <h2>{{ data.title }}</h2>
-    <p>{{ data.content }}</p>
-{% endblock %}
