--- templates/home/index.html.twig
@@ -1,14 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}ブログ{% endblock %}
-
-{% block body %}
-    <h2>記事一覧</h2>
-     <ul>
-         {% for item in data %}
-             <li>
-                 <a href="/home/{{ item.id }}">{{ item.title }}</a>
-             </li>
-         {% endfor %}
-     </ul>
-{% endblock %}
