--- templates/post/index.html.twig
@@ -1,10 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}記事投稿{% endblock %}
-
-{% block body %}
-    <h2>{{ message }}</h2>
-    {{ form_start(form) }}
-    {{ form_widget(form) }}
-    {{ form_end(form) }}
-{% endblock %}
