--- templates/update/update.html.twig
@@ -1,12 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}更新{% endblock %}
-
-{% block body %}
-    <h2>{{ message }}</h2>
-    {{ form_start(form) }}
-    {{ form_widget(form) }}
-    {{ form_end(form) }}
-    <p>記事削除</p>
-    <button class="btn btn-primary"><a href="../delete/{{ id }}" class="text-white">削除</a> </button>
-{% endblock %}
